Added a bit more info to README.
Signed-off-by: David Brown <david.brown@obson.net>

Commit failed using git gui. Undone ('amended') to retry using cola git.

Didn't work, but message said 'Pull first?', so trying that.

Allegedly worked (though no changes) so retrying.

Failed again. Will try Kraken, as that doesn't seem to depend on the installed
version of git and consequently works if you can sort out how to do it. One thing
to remember is always to leave GitKraken in a repository when closing so that it 
can open again in good order. Otherwise it can end up trying to open every file
in the filesystem.